LayoutTests/fast/media/mq-any-pointer-invalid-expected.txt

Test invalid input with the any-pointer media feature.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.matchMedia("(any-pointer:)").matches is false
PASS document.styleSheets[1].cssRules[0].media.length is 1
PASS window.matchMedia("(any-pointer: )").matches is false
PASS document.styleSheets[1].cssRules[0].media.length is 1
PASS window.matchMedia("(any-pointer: WebKit)").matches is false
PASS document.styleSheets[1].cssRules[0].media.length is 1
PASS window.matchMedia("(any-pointer: pointer)").matches is false
PASS document.styleSheets[1].cssRules[0].media.length is 1
PASS window.matchMedia("(any-pointer: hover)").matches is false
PASS document.styleSheets[1].cssRules[0].media.length is 1
PASS window.matchMedia("(any-pointer: portrait)").matches is false
PASS document.styleSheets[1].cssRules[0].media.length is 1
PASS window.matchMedia("(any-pointer: \"None\")").matches is false
PASS document.styleSheets[1].cssRules[0].media.length is 1
PASS window.matchMedia("(any-pointer: 'None')").matches is false
PASS document.styleSheets[1].cssRules[0].media.length is 1
PASS successfullyParsed is true

TEST COMPLETE
